Fix a crash in skia pipeline, when empty reorder block is first
Fix a crash in skia pipeline, which happens if an empty reorder
barrier is inserted at index 0 in a SkiaDisplayList.
Add a unit test that repro the bug (unit test is crashing wihtout
the fix and passing with the fix).

Test: built and ran skia pipeline with angler-eng and run HWUI unit
tests.

+++ libs/hwui/pipeline/skia/ReorderBarrierDrawables.h
@@ -47,8 +47,8 @@ protected:
     virtual void onDraw(SkCanvas* canvas) override;
 
 private:
-    size_t mEndChildIndex;
-    size_t mBeginChildIndex;
+    int mEndChildIndex;
+    int mBeginChildIndex;
     FatVector<RenderNodeDrawable*, 16> mChildren;
     SkiaDisplayList* mDisplayList;
